# Optimize `maybe_macroexpand`